Refactored the UI files Put the screens in screens/press and the widget in widget/tourewidget
