+++ lib/screens/press/press_event_list_page.dart
@@ -1,18 +1,17 @@
+import 'package:allthingscharmaine/screens/press/press_event_detail_page.dart';
 import 'package:allthingscharmaine/utils/custom_colors.dart';
-import 'package:allthingscharmaine/widgets/press_article_item.dart';
+import 'package:allthingscharmaine/widgets/tourewidgets/press_event_item.dart';
 import 'package:flutter/material.dart';
 
-import 'press_article_detail_page.dart';
-
-class PressArticleList extends StatefulWidget{
+class PressEventList extends StatefulWidget{
 
   @override
   State createState() {
-    return _PressArticleListState();
+    return _PressEventListSate();
   }
 }
 
-class _PressArticleListState extends State<PressArticleList>{
+class _PressEventListSate extends State<PressEventList>{
   ScrollController _scrollController =
   ScrollController(initialScrollOffset: 0.0, keepScrollOffset: true);
   List listData;
@@ -30,10 +29,9 @@ class _PressArticleListState extends State<PressArticleList>{
         iconTheme: IconThemeData(
           color: CustomColors.TITLE_COLOR,
         ),
-        automaticallyImplyLeading: true,
         title: Text('press', style: TextStyle(fontWeight: FontWeight.w500, fontFamily: 'Poppins', fontSize:15.0, color: CustomColors.TITLE_COLOR),),
         elevation: 0.0,
-        actions: [
+        actions: <Widget>[
           IconButton(
             icon: Image.asset('assets/notification.png'),
             tooltip: 'Notification',
@@ -49,7 +47,7 @@ class _PressArticleListState extends State<PressArticleList>{
           mainAxisAlignment: MainAxisAlignment.start,
           children: [
             SizedBox(height: 5.0,),
-            Text('articles', style: TextStyle(color: CustomColors.TITLE_COLOR,
+            Text('events', style: TextStyle(color: CustomColors.TITLE_COLOR,
                 fontFamily: 'Poppins', fontSize: 30.0, fontWeight: FontWeight.w600),),
             SizedBox(height: 10.0,),
             Expanded(child: ListView.builder(
@@ -65,8 +63,8 @@ class _PressArticleListState extends State<PressArticleList>{
                         fontFamily: 'Poppins',
                         fontWeight: FontWeight.w400,),),),);
                   }else{
-                    return GestureDetector(child: Container(child: PressArticleItem(listData[index]), margin: EdgeInsets.only(bottom: 15.0),),
-                    onTap: (){Navigator.push(context, MaterialPageRoute(builder: (context) => ArticleDetail(listData[index])));},);
+                    return GestureDetector(child: Container(child: PressEventItem(listData[index]), margin: EdgeInsets.only(bottom: 15.0),),
+                    onTap: (){Navigator.push(context, MaterialPageRoute(builder: (context) => EventDetail(listData[index])));},);
                   }
                 })),
           ],
@@ -77,6 +75,6 @@ class _PressArticleListState extends State<PressArticleList>{
 
   @override
   void initState() {
-    listData = Data.getArticleData();
+    listData = Data.getEventData();
   }
 }
